feat: add support of TLS connection between coordinator and web3signer

+++ coordinator/app/src/main/kotlin/linea/coordinator/config/v2/toml/MessageAnchoringConfigToml.kt
@@ -92,8 +92,8 @@ data class MessageAnchoringConfigToml(
         ethLogsSearchBlockChunkSize = l1EventScraping.ethLogsSearchBlockChunkSize,
       ),
       anchoringTickInterval = anchoringTickInterval,
-      messageQueueCapacity = messageQueueCapacity.toUInt(),
-      maxMessagesToAnchorPerL2Transaction = maxMessagesToAnchorPerL2Transaction.toUInt(),
+      messageQueueCapacity = messageQueueCapacity,
+      maxMessagesToAnchorPerL2Transaction = maxMessagesToAnchorPerL2Transaction,
       signer = signer.reified(),
       gas = MessageAnchoringConfig.GasConfig(
         maxFeePerGasCap = gas.maxFeePerGasCap,
